Implement Page Object Model pattern (#192)
This pull request includes changes to the end-to-end tests of the
`DivingCalculator.Web` project. The changes primarily focus on
refactoring the tests to improve readability and maintainability. The
tests have been restructured to use Page Object Model (POM), which is a
design pattern that helps in reducing code duplication and improving
test maintenance.