exclude dist from codeql (#235)
This pull request primarily focuses on the implementation of CodeQL code
scanning for the project. CodeQL is a tool by GitHub that helps discover
potential vulnerabilities in the codebase. Two significant changes were
made: the addition of a custom CodeQL configuration file and the
modification of the continuous integration (CI) workflow file to use
this custom configuration.