Flatten dir structure (#253)
This pull request includes changes primarily focused on refactoring the
project structure. The main change is the removal of the
`DivingCalculator.Web` directory, and the subsequent update of all paths
in various configuration files to reflect this change. The changes span
across several files including configuration files for the development
container, CodeQL, Dependabot, GitHub workflows, and Visual Studio Code.